+++ example/lib/src/view/home/page_01.dart
@@ -225,4 +225,15 @@ class Page1State extends StateX<Page1> with EventsStateMixin<Page1> {
       }
     }
   }
+
+  /// Catch it if the initAsync() throws an error
+  /// The FutureBuilder will fail, but you can examine the error
+  @override
+  void catchAsyncError(Object error) {
+    final con = Controller();
+    if (con.errorCatchAsyncError) {
+      con.errorCatchAsyncError = false;
+      throw AssertionError('Error in Catch!');
+    }
+  }
 }